Update: currently only aims for flushes
Currently aims for getting a flush in each of the first four columns,
with extras being placed in the fifth. If the extra column gets filled,
extra cards will then be placed in the first non-full column (Hoping to
improve this, maybe random?). Average mean of 100 games is around 80
points, which isn't terrible, but can be improved.