Minor improvements
Added batch testing and cleaned up the unused code. Also make it so that
if the extra cards column is full, the card will be placed in the column
with the most free space. The idea here being that if a column almost
has a flush, then it would be stupid to ruin that by placing an extra
card there.